requirements.txt
asciichartpy==1.5.25
confluent_kafka==2.13.0
numpy==2.4.1
python_binance==1.0.34
ta_lib==0.6.8
websocket_client==1.9.0
